Move RequestBase abstract model to core app
